refactor: split main_menu.dart into part files under size cap (#3594)
Extract the main-menu button, footer, and pixelArt scroll-bracket widget
groups out of the 1025-line `main_menu.dart` host into three sibling
`part of 'main_menu.dart';` files, mirroring the repo's existing
trade-screen part-file pattern. Behavior, public API, and private
visibility are unchanged; the host file drops to 493 physical lines,
satisfying the >1000-line split rule and the issue's <=700 target.

Refs #3594
